rename to client_test
and remove doc.go

There seems to be a bug with go, if I include a doc.go it complains about the package name being mismatched (integration vs integration_test).
