--- integration_test/doc.go
@@ -1,2 +1,0 @@
-// Package integration_test contains helpers and integration tests for go-nitro clients
-package integration_test // import "github.com/statechannels/go-nitro/integration_test"
